[Frontend] - refactoring only. Transferring some functions from mapEvents.js to mapMenuActions.js, etc.
