+++ src/main/java/com/dungeoncrawler/Javiarenka/dungeonMapGenerator/webStuff/debug/mapStyle.css
@@ -86,8 +86,8 @@
 
 .stage-grid{
 	display: grid;
-	grid-template-columns: repeat(30, 1fr);
-	grid-template-rows: repeat(30, 1fr);
+	grid-template-columns: repeat(35, 1fr);
+	grid-template-rows: repeat(25, 1fr);
 	grid-column-gap: 0;
 	grid-row-gap: 0;
 	height: 90vh;
