+++ src/main/resources/static/css/mapStyleToolbar.css
@@ -1,3 +1,5 @@
+/* TODO: potworzyć zmienne z kolorami (na bazie rgb lub hex) i podmienić je w kodzie */
+
 :root {
     --toolbar-height: 15vh;
     --toolbar-padding: 1vh;
